# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LogForge API",
    "description": "\nLogForge API provides a robust and scalable log management and ingestion platform.\nIt supports multiple ingestion protocols (HTTP, GELF UDP) and provides real-time log monitoring through WebSockets.\n\n### Features:\n* **Log Ingestion**: High-performance log capture using Elasticsearch/OpenSearch.\n* **Organization**: Manage Projects and Channels with granular permissions.\n* **Real-time**: Live log streaming via WebSockets.\n* **Security**: JWT-based authentication with role-based access control (RBAC).\n* **Multi-tenant**: Secure isolated environments per project.\n",
    "contact": {
      "name": "LogForge Support",
      "url": "https://github.com/loickadjiwanou/log-forge"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.7"
  },
  "paths": {
...
    "/api/projects/": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Projects",
        "description": "Retrieve a paginated list of projects accessible by the current user.",
        "operationId": "list_projects_api_projects__get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 99,
              "title": "Size"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create Project",
        "description": "Create a new project. Requires administrator privileges.",
        "operationId": "create_project_api_projects__post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/channels/": {
      "get": {
        "tags": [
          "channels"
        ],
        "summary": "List Channels",
        "description": "Retrieve a paginated list of channels. Optionally filtered by project ID.",
        "operationId": "list_channels_api_channels__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 99,
              "title": "Size"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "channels"
        ],
        "summary": "Create Channel",
        "description": "Create a new channel within a specific project.",
        "operationId": "create_channel_api_channels__post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChannelCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/logs": {
      "get": {
        "tags": [
          "logs"
        ],
        "summary": "Search Logs",
        "description": "Search and filter logs with pagination. Supports full-text search and faceted filtering.",
        "operationId": "list_logs_api_logs_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter by project ID",
              "title": "Project Id"
            },
            "description": "Filter by project ID"
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter by log level (info, error, etc)",
              "title": "Level"
            },
            "description": "Filter by log level (info, error, etc)"
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter by channel name",
              "title": "Channel"
            },
            "description": "Filter by channel name"
          },
          {
            "name": "environment",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter by environment (production, staging, etc)",
              "title": "Environment"
            },
            "description": "Filter by environment (production, staging, etc)"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Full-text search query",
              "title": "Search"
            },
            "description": "Full-text search query"
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "ISO timestamp start range",
              "title": "Date From"
            },
            "description": "ISO timestamp start range"
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "ISO timestamp end range",
              "title": "Date To"
            },
            "description": "ISO timestamp end range"
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Comma-separated list of tags",
              "title": "Tags"
            },
            "description": "Comma-separated list of tags"
          },
          {
            "name": "grouped_hash",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter by error group hash",
              "title": "Grouped Hash"
            },
            "description": "Filter by error group hash"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 99,
              "title": "Size"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/settings/": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Global Settings",
        "description": "Retrieve the current user theme and the global SMTP configuration.",
        "operationId": "get_settings_api_settings__get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/setup/admin": {
      "post": {
        "summary": "Setup Admin",
        "operationId": "setup_admin_api_setup_admin_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupAdminRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ChannelCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "project_id"
        ],
        "title": "ChannelCreate"
      },
      "ChannelListResponse": {
        "properties": {
          "channels": {
            "items": {
              "$ref": "#/components/schemas/ChannelResponse"
            },
            "type": "array",
            "title": "Channels"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "channels",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "ChannelListResponse"
      },
      "ChannelResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "project_id",
          "created_at"
        ],
        "title": "ChannelResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LogListResponse": {
        "properties": {
          "logs": {
            "items": {
              "$ref": "#/components/schemas/LogResponse"
            },
            "type": "array",
            "title": "Logs"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          }
        },
        "type": "object",
        "required": [
          "logs",
          "total",
          "page",
          "size"
        ],
        "title": "LogListResponse"
      },
      "LogResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "level": {
            "type": "string",
            "title": "Level"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "channel": {
            "type": "string",
            "title": "Channel"
          },
          "environment": {
            "type": "string",
            "title": "Environment"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "project_name": {
            "type": "string",
            "title": "Project Name"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "default": {}
          },
          "stack_trace": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stack Trace"
          },
          "user_info": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Info"
          },
          "device_info": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Info"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "default": []
          },
          "has_replay": {
            "type": "boolean",
            "title": "Has Replay",
            "default": false
          },
          "grouped_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grouped Hash"
          },
          "alert_email_sent": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alert Email Sent"
          },
          "ingest_protocol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingest Protocol"
          },
          "gelf_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gelf Version"
          },
          "source_host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Host"
          },
          "syslog_level": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Syslog Level"
          }
        },
        "type": "object",
        "required": [
          "id",
          "level",
          "message",
          "channel",
          "environment",
          "project_name",
          "timestamp"
        ],
        "title": "LogResponse"
      },
...
      "ProjectCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "environment": {
            "type": "string",
            "title": "Environment",
            "default": "production"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ProjectCreate"
      },
      "ProjectListResponse": {
        "properties": {
          "projects": {
            "items": {
              "$ref": "#/components/schemas/ProjectResponse"
            },
            "type": "array",
            "title": "Projects"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "projects",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "ProjectListResponse"
      },
      "ProjectResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "environment": {
            "type": "string",
            "title": "Environment"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "environment",
          "api_key",
          "user_id",
          "created_at"
        ],
        "title": "ProjectResponse"
      },
...
      "SMTPConfigResponse": {
        "properties": {
          "host": {
            "type": "string",
            "title": "Host"
          },
          "port": {
            "type": "integer",
            "title": "Port"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "from_email": {
            "type": "string",
            "title": "From Email"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          }
        },
        "type": "object",
        "required": [
          "host",
          "port",
          "username",
          "from_email",
          "language",
          "enabled"
        ],
        "title": "SMTPConfigResponse"
      },
...
      "SettingsResponse": {
        "properties": {
          "theme": {
            "type": "string",
            "title": "Theme"
          },
          "smtp_config": {
            "$ref": "#/components/schemas/SMTPConfigResponse"
          }
        },
        "type": "object",
        "required": [
          "theme",
          "smtp_config"
        ],
        "title": "SettingsResponse"
      },
      "SetupAdminRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "app_name": {
            "type": "string",
            "title": "App Name",
            "default": "LogForge"
          },
          "primary_color": {
            "type": "string",
            "title": "Primary Color",
            "default": "#10b981"
          },
          "theme": {
            "type": "string",
            "title": "Theme",
            "default": "dark"
          }
        },
        "type": "object",
        "required": [
          "name",
          "email",
          "password"
        ],
        "title": "SetupAdminRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}